requirements.txt
<<<<<<< HEAD
# ✅ FastAPI & Server
fastapi
uvicorn
gunicorn

# ✅ Environment & Request Handling
pydantic
requests
python-dotenv
asyncio

# ✅ Google API (For Google Drive Integration)
=======
fastapi
uvicorn
gunicorn
pydantic
requests
python-dotenv
>>>>>>> heroku/main
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
<<<<<<< HEAD

# ✅ Document Processing
=======
>>>>>>> heroku/main
pdfplumber
pytesseract
pillow
pdf2image
opencv-python
PyMuPDF
python-docx
python-pptx
pandas
<<<<<<< HEAD
openpyxl

# ✅ LangChain & OpenAI
langchain-community  # (Avoid using 'langchain' monolithic package)
langchain-pinecone
langchain-openai
openai
tiktoken  # (For OpenAI Token Counting)

# ✅ Pinecone Integration
pinecone-client

# ✅ LlamaIndex (New Modular Structure)
llama-index-core
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-vector-stores-pinecone

# ✅ NLP & Machine Learning
scikit-learn  # (For text similarity search)
nltk  # (For NLP preprocessing)
spacy  # (For advanced text processing)

# ✅ Text Evaluation
rouge-score

# ✅ Performance Optimization
numexpr  # (For fast Pandas computations)
=======
langchain
langchain-community
langchain-pinecone
openai
langchain-openai
pinecone-client
rouge-score
>>>>>>> heroku/main
